refactor(v1): freeze contracts and isolate ownership
Extract Kubernetes ownership and destructive preflight from lifecycle orchestration without changing behavior. Freeze the complete v1alpha1 schema and fixture inventory behind digest validation.\n\nRefs: #236
