+++ cli/contracts/v1alpha1/README.md
@@ -16,3 +16,12 @@ They are not backend configuration and must not be rendered or applied.
 `DisposableClusterEvidence` is release-only attestation generated by the live
 Kubernetes producer. It binds the exact candidate commit and tested binary
 digest, and records each required lifecycle scenario only after it passes.
+
+## Contract freeze
+
+`contract-freeze.sha256` is the reviewed inventory of every public schema and
+conformance fixture in this version. Tests fail when a file is added, removed,
+or changed without updating that manifest. Updating it is an explicit public
+API change: the review must assess compatibility, update the relevant valid and
+invalid fixtures, and use a new contract version when the change would
+reinterpret or reject previously valid input.
