+++ cli/contracts/v1alpha1/contract-freeze.sha256
@@ -1,0 +1,26 @@
+d8f810ee748caafa0fb1ac7396e5672a86c29c2d8e8a5ac6eb4ee4d4731ea699  fixtures/invalid/compiler-duplicate-service.json
+6ea7d8f8af2727aeafd335a3e595f3c44c665433ae054a60f100abd57bbd6c57  fixtures/invalid/deployment-target-raw-credential.json
+8932f6d94f680137a710508963ab19baf010deed0634bf37248513c25b524d55  fixtures/invalid/disposable-cluster-evidence-duplicate-scenario.json
+9a7c69f7cf2e17e532a9c3a8a03442a7718282062a11f2947909fe1f674bff7d  fixtures/invalid/disposable-cluster-evidence-missing-scenario.json
+f1e1359f173048a75acdfa08aa6a8916cfd5a261c632a64efcd90fe0880c6918  fixtures/invalid/disposable-cluster-evidence-unexecuted-scenario.json
+14e4ae729f129bb6ab6001c44c94cfcdf67c2182415cbf928196d8c88d5a9fbd  fixtures/invalid/disposable-cluster-evidence-unknown-scenario.json
+8faa5e2c6a0d392311750a196577b3f4471dfbe9027d8435df0c6707856a9f1c  fixtures/invalid/release-gate-evidence-missing-artifacts.json
+23fd6161ef146edf51e14d1fb642ff9c92abc51063defb64aa71d28b74ef96e2  fixtures/invalid/workspace-image-spec-unknown-version.json
+42531855b53e4d3d23e6e1bcc73f0f9c94e8aa4e6cfa61efeb317a16963901da  fixtures/valid/backend-descriptor.json
+75a441b48a0a2c8570861d7d080686403403c3987581a0a4b152baaa2fe3fa38  fixtures/valid/compose-plan.yaml
+1f4b5b9c2fe6ae477cc59818c19678312fad46d1d2ccb65112e854a5aaa47ccf  fixtures/valid/connection-target.json
+5dbe97730b604a91a0d27b938664e4f7efe98123a0ddb5b9211357bf7829a66c  fixtures/valid/deployment-record.json
+2c2e55f13c9a5ee8292df4e22da58fc4f2a872e433ff3fe8018474dd165b9049  fixtures/valid/deployment-target.json
+c2f12c9f050126b4e6424ac0e3e156c327dd005585cd67be43d638a7881e8352  fixtures/valid/disposable-cluster-evidence.json
+960a57eb2494b9a0fc1138dd007f8e24395d7583ee6847395243b2b6ada5c399  fixtures/valid/kubernetes-plan.yaml
+068af9078bce7f1e908df62883cdcc3cc657ef9e30e4ad3166ac1710a081d295  fixtures/valid/release-gate-evidence.json
+caf2a3e16b2d6bf189323561daed3e4dcbe34bcc07c215f4d7ba343a79d50a18  fixtures/valid/workspace-fleet-spec.json
+030e0b75eae7b3171ea27065f15bca924d4c4a6bb6c127e23cfd81df1ec0cc39  fixtures/valid/workspace-image-spec.json
+727284fe4456f34493565bdeb91df635dfb96fc666304c25711c4e8903b6c98c  schemas/backend-descriptor.schema.json
+7bc385836695f43da95febe8ec8a5e47a813e312e0e4fea5567a9498dbf6baf3  schemas/connection-target.schema.json
+3532f0caeb27a896a37ccb0b72d12f9e3ec584a938ebe46226f7717f8d0884cf  schemas/deployment-record.schema.json
+b12600f442f07be91be0ea2ef967d6bdf1ea334cf41e648a5bd0d9fa7acddaa1  schemas/deployment-target.schema.json
+44d81f6ac7a30234d331bd8efed3d27619b1220d0614eb710dbd99c45c3e42be  schemas/disposable-cluster-evidence.schema.json
+1591ad63ed75e9add616438e5f4f2242c4c8b0f1365b425819ace030f1991e0b  schemas/release-gate-evidence.schema.json
+88e53c4a6eae76605f1ee2b5011c33b94621007219b1aa4fe4cbd881f287e449  schemas/workspace-fleet-spec.schema.json
+4052f2f259aa04856b35230a29e5c20725ed17e281b61379020fb5ad3baaaa75  schemas/workspace-image-spec.schema.json
